Frivolous Gravias Live session 3 supporting docs
A version with some comments from the live coding demo, and another with more comments and mental notes jotted down during production of the recording being uploaded to youTube under the channel for Frivolous Gravitas. I'm including 2 files since they are the same subject matter but the comments pertain specifically to some tangents taken during the live recording. They're essentially the same, but some of the tangents I took in the video may not be in the notes file, and we took some fun tangents in this session so might as well include both, I figured.