uploaded canvas sketch of field lines w/ animation
An animated sketch up of javascript and html canvas, this time mapping the field using for loops to build an object which stores index lookup either by row or by column. A bunch of commented out options explored and some notes on effects those changes make. Field lines in this sketch can be used to displace or distort other maps like images or textures or particle flows in practise.